+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/dao/preset/DriverInfoPresetCriteriaMaker.java
@@ -82,7 +82,6 @@ public class DriverInfoPresetCriteriaMaker implements PresetCriteriaMaker {
         }
     }
 
-
     private List<Long> longList(List<LongIdKey> list) {
         return list.stream().map(LongIdKey::getLongId).collect(Collectors.toList());
     }