+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/driver/CronDriverRegistry.java
@@ -1,6 +1,5 @@
 package com.dwarfeng.judge.impl.handler.driver;
 
-import com.dwarfeng.judge.impl.handler.DriverProvider;
 import com.dwarfeng.judge.stack.bean.entity.DriverInfo;
 import com.dwarfeng.judge.stack.exception.DriverException;
 import com.dwarfeng.judge.stack.handler.Driver;
@@ -10,41 +9,55 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
+import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Component;
 
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
- * 固定间隔驱动提供器。
+ * Cron驱动提注册。
  *
  * @author DwArFeng
- * @since beta-1.0.0
+ * @since 1.3.0
  */
 @Component
-public class FixedRateDriverProvider implements DriverProvider {
+public class CronDriverRegistry extends AbstractDriverRegistry {
 
-    public static final String SUPPORT_TYPE = "fixed_rate_driver";
+    public static final String DRIVER_TYPE = "cron_driver";
 
     @Autowired
-    private FixedRateDriver fixedRateDriver;
+    private CronDriver cronDriver;
+
+    public CronDriverRegistry() {
+        super(DRIVER_TYPE);
+    }
 
     @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public String provideLabel() {
+        return "Cron驱动器";
+    }
+
+    @Override
+    public String provideDescription() {
+        return "根据指定的Cron表达式定时驱动的驱动器";
+    }
+
+    @Override
+    public String provideExampleContent() {
+        return "0/2 * * * * *";
     }
 
     @Override
     public Driver provide() {
-        return fixedRateDriver;
+        return cronDriver;
     }
 
     @Component
-    public static class FixedRateDriver implements Driver {
+    public static class CronDriver implements Driver {
 
         @Autowired
         private ThreadPoolTaskScheduler scheduler;
@@ -53,21 +66,21 @@ public class FixedRateDriverProvider implements DriverProvider {
 
         private final Lock lock = new ReentrantLock();
         private final Set<ScheduledFuture<?>> scheduledFutures = new HashSet<>();
-        private final Set<FixedRateProcessor> fixedRateProcessors = new HashSet<>();
+        private final Set<CronProcessor> cronProcessors = new HashSet<>();
 
         @Override
         public void register(DriverInfo driverInfo) throws DriverException {
             lock.lock();
             try {
                 LongIdKey sectionKey = driverInfo.getSectionKey();
-                long rate = Long.parseLong(driverInfo.getContent());
-                FixedRateProcessor fixedRateProcessor = new FixedRateProcessor(
+                String cron = driverInfo.getContent();
+                CronProcessor cronProcessor = new CronProcessor(
                         evaluateService,
                         sectionKey
                 );
-                ScheduledFuture<?> scheduledFuture =
-                        scheduler.scheduleAtFixedRate(fixedRateProcessor, rate);
-                fixedRateProcessors.add(fixedRateProcessor);
+                CronTrigger cronTrigger = new CronTrigger(cron);
+                ScheduledFuture<?> scheduledFuture = scheduler.schedule(cronProcessor, cronTrigger);
+                cronProcessors.add(cronProcessor);
                 scheduledFutures.add(scheduledFuture);
             } catch (Exception e) {
                 throw new DriverException(e);
@@ -83,8 +96,8 @@ public class FixedRateDriverProvider implements DriverProvider {
                 for (ScheduledFuture<?> scheduledFuture : scheduledFutures) {
                     scheduledFuture.cancel(true);
                 }
-                for (FixedRateProcessor fixedRateProcessor : fixedRateProcessors) {
-                    fixedRateProcessor.shutdown();
+                for (CronProcessor cronProcessor : cronProcessors) {
+                    cronProcessor.shutdown();
                 }
             } finally {
                 lock.unlock();
@@ -92,9 +105,9 @@ public class FixedRateDriverProvider implements DriverProvider {
         }
     }
 
-    private static class FixedRateProcessor implements Runnable {
+    private static class CronProcessor implements Runnable {
 
-        private static final Logger LOGGER = LoggerFactory.getLogger(FixedRateProcessor.class);
+        private static final Logger LOGGER = LoggerFactory.getLogger(CronProcessor.class);
 
         private final EvaluateService evaluateService;
         private final LongIdKey sectionKey;
@@ -102,7 +115,7 @@ public class FixedRateDriverProvider implements DriverProvider {
         private final Lock lock = new ReentrantLock();
         private boolean runningFlag = true;
 
-        private FixedRateProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
+        private CronProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
             this.evaluateService = evaluateService;
             this.sectionKey = sectionKey;
         }
@@ -115,7 +128,7 @@ public class FixedRateDriverProvider implements DriverProvider {
                     return;
                 }
 
-                LOGGER.debug("计划时间已到达, fixed rate 驱动器驱动 " + sectionKey + " 部件执行评估动作...");
+                LOGGER.debug("计划时间已到达, cron驱动器驱动 " + sectionKey + " 部件执行评估动作...");
                 evaluateService.evaluate(sectionKey);
             } catch (Exception e) {
                 LOGGER.warn("记录 " + sectionKey + " 时出现异常, 放弃本次记录", e);