+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/handler/driver/FixedRateDriverRegistry.java
@@ -1,6 +1,5 @@
 package com.dwarfeng.judge.impl.handler.driver;
 
-import com.dwarfeng.judge.impl.handler.DriverProvider;
 import com.dwarfeng.judge.stack.bean.entity.DriverInfo;
 import com.dwarfeng.judge.stack.exception.DriverException;
 import com.dwarfeng.judge.stack.handler.Driver;
@@ -10,42 +9,54 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
-import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Component;
 
 import java.util.HashSet;
-import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
 
 /**
- * Cron驱动提供器。
+ * 固定间隔驱动提供器。
  *
  * @author DwArFeng
- * @since beta-1.0.0
+ * @since 1.3.0
  */
 @Component
-public class CronDriverProvider implements DriverProvider {
+public class FixedRateDriverRegistry extends AbstractDriverRegistry {
 
-    public static final String SUPPORT_TYPE = "cron_driver";
+    public static final String DRIVER_TYPE = "fixed_rate_driver";
 
     @Autowired
-    private CronDriver cronDriver;
+    private FixedRateDriver fixedRateDriver;
+
+    public FixedRateDriverRegistry() {
+        super(DRIVER_TYPE);
+    }
 
     @Override
-    public boolean supportType(String type) {
-        return Objects.equals(SUPPORT_TYPE, type);
+    public String provideLabel() {
+        return "固定频率驱动器";
+    }
+
+    @Override
+    public String provideDescription() {
+        return "根据指定的间隔定时驱动，如果某一次驱动晚于间隔，则后续驱动的时间会提前，以保持频率不变。";
+    }
+
+    @Override
+    public String provideExampleContent() {
+        return "60000";
     }
 
     @Override
     public Driver provide() {
-        return cronDriver;
+        return fixedRateDriver;
     }
 
     @Component
-    public static class CronDriver implements Driver {
+    public static class FixedRateDriver implements Driver {
 
         @Autowired
         private ThreadPoolTaskScheduler scheduler;
@@ -54,21 +65,21 @@ public class CronDriverProvider implements DriverProvider {
 
         private final Lock lock = new ReentrantLock();
         private final Set<ScheduledFuture<?>> scheduledFutures = new HashSet<>();
-        private final Set<CronProcessor> cronProcessors = new HashSet<>();
+        private final Set<FixedRateProcessor> fixedRateProcessors = new HashSet<>();
 
         @Override
         public void register(DriverInfo driverInfo) throws DriverException {
             lock.lock();
             try {
                 LongIdKey sectionKey = driverInfo.getSectionKey();
-                String cron = driverInfo.getContent();
-                CronProcessor cronProcessor = new CronProcessor(
+                long rate = Long.parseLong(driverInfo.getContent());
+                FixedRateProcessor fixedRateProcessor = new FixedRateProcessor(
                         evaluateService,
                         sectionKey
                 );
-                CronTrigger cronTrigger = new CronTrigger(cron);
-                ScheduledFuture<?> scheduledFuture = scheduler.schedule(cronProcessor, cronTrigger);
-                cronProcessors.add(cronProcessor);
+                ScheduledFuture<?> scheduledFuture =
+                        scheduler.scheduleAtFixedRate(fixedRateProcessor, rate);
+                fixedRateProcessors.add(fixedRateProcessor);
                 scheduledFutures.add(scheduledFuture);
             } catch (Exception e) {
                 throw new DriverException(e);
@@ -84,8 +95,8 @@ public class CronDriverProvider implements DriverProvider {
                 for (ScheduledFuture<?> scheduledFuture : scheduledFutures) {
                     scheduledFuture.cancel(true);
                 }
-                for (CronProcessor cronProcessor : cronProcessors) {
-                    cronProcessor.shutdown();
+                for (FixedRateProcessor fixedRateProcessor : fixedRateProcessors) {
+                    fixedRateProcessor.shutdown();
                 }
             } finally {
                 lock.unlock();
@@ -93,9 +104,9 @@ public class CronDriverProvider implements DriverProvider {
         }
     }
 
-    private static class CronProcessor implements Runnable {
+    private static class FixedRateProcessor implements Runnable {
 
-        private static final Logger LOGGER = LoggerFactory.getLogger(CronProcessor.class);
+        private static final Logger LOGGER = LoggerFactory.getLogger(FixedRateProcessor.class);
 
         private final EvaluateService evaluateService;
         private final LongIdKey sectionKey;
@@ -103,7 +114,7 @@ public class CronDriverProvider implements DriverProvider {
         private final Lock lock = new ReentrantLock();
         private boolean runningFlag = true;
 
-        private CronProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
+        private FixedRateProcessor(EvaluateService evaluateService, LongIdKey sectionKey) {
             this.evaluateService = evaluateService;
             this.sectionKey = sectionKey;
         }
@@ -116,7 +127,7 @@ public class CronDriverProvider implements DriverProvider {
                     return;
                 }
 
-                LOGGER.debug("计划时间已到达, cron驱动器驱动 " + sectionKey + " 部件执行评估动作...");
+                LOGGER.debug("计划时间已到达, fixed rate 驱动器驱动 " + sectionKey + " 部件执行评估动作...");
                 evaluateService.evaluate(sectionKey);
             } catch (Exception e) {
                 LOGGER.warn("记录 " + sectionKey + " 时出现异常, 放弃本次记录", e);